requirements.txt
steamclient